logs: cast uint32_t-typed args printed with %u / %d
picolibc on xtensa (IDF v6) defines uint32_t as unsigned long, so
every %u or %d of a uint32_t (ah_data_idx_t, lejp ctx->line, the
jose map lengths, alt-svc ma_secs, lws_dll2_count()) is a
-Werror=format failure there.  Found by building natively with
uint32_t forced to unsigned long through a shim header; these are all
the sites in lib/, including ones not built on esp32.

+++ lib/jose/jwe/enc/aeskw.c
@@ -143,7 +143,7 @@ lws_jwe_auth_and_decrypt_aeskw_cbc_hs(struct lws_jwe *jwe)
 					jwe->jose.enc_alg->hmac_type) +
 			LWS_JWE_RFC3394_OVERHEAD_BYTES) {
 		lwsl_err("%s: bad EKEY len %u\n", __func__,
-			 jwe->jws.map.len[LJWE_EKEY]);
+			 (unsigned int)jwe->jws.map.len[LJWE_EKEY]);
 
 		return -1;
 	}
